moves glyph/sum_channels to feature package
