+++ docs/source/api/menpo/feature/glyph.rst
@@ -1,0 +1,7 @@
+.. _menpo-feature-visualize-glyph:
+
+.. currentmodule:: menpo.feature.visualize
+
+glyph
+=====
+.. autofunction:: glyph
