+++ docs/source/api/menpo/feature/sum_channels.rst
@@ -1,0 +1,7 @@
+.. _menpo-feature-visualize-sum_channels:
+
+.. currentmodule:: menpo.feature.visualize
+
+sum_channels
+============
+.. autofunction:: sum_channels
