+++ docs/xref_map.py
@@ -12,6 +12,7 @@ xref_map = {
 'DiscreteAffine': ('class', 'menpo.transform.DiscreteAffine'),
 'from_vector_inplace': ('function', 'menpo.base.Vectorizable.from_vector_inplace'),
 'from_vector': ('function', 'menpo.base.Vectorizable.from_vector'),
+'glyph': ('function', 'menpo.feature.visualize.glyph'),
 'GraphOptionsWidget': ('class', 'menpo.visualize.widgets.GraphOptionsWidget'),
 'GraphPlotter': ('class', 'menpo.visualize.GraphPlotter'),
 'Homogeneous': ('class', 'menpo.transform.Homogeneous'),
@@ -51,6 +52,7 @@ xref_map = {
 'SaveFigureOptionsWidget': ('class', 'menpo.visualize.widgets.SaveFigureOptionsWidget'),
 'Scale': ('class', 'menpo.shape.Scale'),
 'Similarity': ('class', 'menpo.transform.Similarity'),
+'sum_channels': ('function', 'menpo.feature.visualize.sum_channels'),
 'Targetable': ('class', 'menpo.base.Targetable'),
 'TextPrintWidget': ('class', 'menpo.visualize.widgets.TextPrintWidget'),
 'Transform': ('class', 'menpo.transform.Transform'),
